+++ opendaw_mcp/__init__.py
@@ -27,6 +27,7 @@ from .utils import (
     _safe_path,
     _clamp_script_param,
     _detect_bpm,
+    _detect_key,
 )
 from .music_theory import (
     NOTE_TO_PITCH,
@@ -58,6 +59,7 @@ __all__ = [
     "_safe_path",
     "_clamp_script_param",
     "_detect_bpm",
+    "_detect_key",
     "NOTE_TO_PITCH",
     "CHORD_INTERVALS",
     "SCALE_INTERVALS",